ResponseEntity를 이용한 HttpStatus 처리 구현

+++ src/main/kotlin/com/mircoservices/study/controller/CustomerController.kt
@@ -3,6 +3,8 @@ package com.mircoservices.study.controller
 import com.mircoservices.study.Interface.CustomerService
 import com.mircoservices.study.data.Customer
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import java.util.concurrent.ConcurrentHashMap
 
@@ -12,27 +14,45 @@ class CustomerController {
     @Autowired
     private lateinit var customerService: CustomerService
 
-    @RequestMapping(value = "/customer/{id}", method = arrayOf(RequestMethod.GET))
-    fun getCustomer(@PathVariable("id") id : Int) = customerService.getCustomer(id)
+    @GetMapping("/customer/{id}")
+    fun getCustomer(@PathVariable("id") id: Int): ResponseEntity<Customer?> {
+        val customer = customerService.getCustomer(id)
+        val status = if (customer == null) HttpStatus.NOT_FOUND else HttpStatus.OK
 
-    @RequestMapping(value = "/customer/", method = arrayOf(RequestMethod.POST))
-    fun postCustomer(@RequestBody customer: Customer) {
+        return ResponseEntity(customer, status)
+    }
+
+    @PostMapping("/customer/")
+    fun postCustomer(@RequestBody customer: Customer) : ResponseEntity<Unit> {
         customerService.createCustomer(customer)
+        return ResponseEntity(Unit, HttpStatus.CREATED)
+        //return ResponseEntity<Unit?>(null, HttpStatus.CREATED)
+
     }
 
-    @RequestMapping(value ="/customer/{id}", method = arrayOf(RequestMethod.DELETE))
-    fun deleteCustomer(@PathVariable("id") id : Int) {
-        customerService.deleteCustomer(id)
+
+    @DeleteMapping("/customer/{id}")
+    fun deleteCustomer(@PathVariable("id") id: Int) : ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if(customerService.getCustomer(id) != null){
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
-    @RequestMapping(value = "/customer/{id}", method = arrayOf(RequestMethod.PUT))
-    fun putCustomer(@PathVariable("id")id: Int, @RequestBody customer: Customer) {
-        customerService.updateCustomer(id, customer)
-    }
+    @PutMapping("/customer/{id}")
+    fun putCustomer(@PathVariable("id") id: Int, @RequestBody customer: Customer) : ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null) {
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.OK
+        }
 
-    @RequestMapping(value = "/customers", method = arrayOf(RequestMethod.GET))
+        return ResponseEntity(Unit, status)
+    }
+    @GetMapping("/customers")
     fun getCustomers(@RequestParam(required = false, defaultValue = "") nameFilter: String) =
             customerService.serachCustomers(nameFilter)
-
-
 }
